# perf(es/minifier): Cache the list of bindings for DCE